refactor(#792): split rig_tests into base + preset write-back suite (sweep)
rig_tests.rs 670 -> 381, + rig_writeback_tests.rs 295 (both under cap; off debt).
The #436 write-back / structural-swap tests move to the sibling; input/project_with/
core_block become pub(super), imported via super::rig_tests. project tests (17) pass,
zero warnings.
