+++ scripts/validate.sh
@@ -74,7 +74,6 @@ crates/infra-cpal/tests/issue_670_cab_swap.rs
 crates/infra-yaml/src/lib_tests.rs
 crates/lv2/src/host.rs
 crates/project/src/block_tests.rs
-crates/project/src/rig_tests.rs
 crates/adapter-gui/ui/app-window.slint
 crates/adapter-gui/ui/desktop_main.slint
 crates/adapter-gui/ui/pages/block_panel_editor.slint
